+++ doc/conf.py
@@ -145,6 +145,7 @@ html_theme_options = {
     "use_edit_page_button": False,
     "navigation_with_keys": False,
     "show_toc_level": 1,
+    "header_links_before_dropdown": 6,
     "navbar_end": ["theme-switcher", "version-switcher", "navbar-icon-links"],
     "analytics": dict(google_analytics_id="G-C8SH9E98QC"),
     "switcher": {